migrate from custom component to core integration
