--- custom_components/open_router/__init__.py
@@ -1,60 +1,0 @@
-"""The OpenRouter integration."""
-
-from __future__ import annotations
-
-import logging
-
-from openai import AsyncOpenAI, AuthenticationError, OpenAIError
-
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_API_KEY, Platform
-from homeassistant.core import HomeAssistant
-from homeassistant.exceptions import ConfigEntryError, ConfigEntryNotReady
-from homeassistant.helpers.httpx_client import get_async_client
-
-_LOGGER = logging.getLogger(__name__)
-
-PLATFORMS = [Platform.AI_TASK, Platform.CONVERSATION]
-
-type OpenRouterConfigEntry = ConfigEntry[AsyncOpenAI]
-
-
-async def async_setup_entry(hass: HomeAssistant, entry: OpenRouterConfigEntry) -> bool:
-    """Set up OpenRouter from a config entry."""
-    client = AsyncOpenAI(
-        base_url="https://openrouter.ai/api/v1",
-        api_key=entry.data[CONF_API_KEY],
-        http_client=get_async_client(hass),
-    )
-
-    # Cache current platform data which gets added to each request (caching done by library)
-    _ = await hass.async_add_executor_job(client.platform_headers)
-
-    try:
-        async for _ in client.with_options(timeout=10.0).models.list():
-            break
-    except AuthenticationError as err:
-        _LOGGER.error("Invalid API key: %s", err)
-        raise ConfigEntryError("Invalid API key") from err
-    except OpenAIError as err:
-        raise ConfigEntryNotReady(err) from err
-
-    entry.runtime_data = client
-
-    await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
-
-    entry.async_on_unload(entry.add_update_listener(_async_update_listener))
-
-    return True
-
-
-async def _async_update_listener(
-    hass: HomeAssistant, entry: OpenRouterConfigEntry
-) -> None:
-    """Handle update."""
-    await hass.config_entries.async_reload(entry.entry_id)
-
-
-async def async_unload_entry(hass: HomeAssistant, entry: OpenRouterConfigEntry) -> bool:
-    """Unload OpenRouter."""
-    return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
