--- custom_components/open_router/manifest.json
@@ -1,14 +1,0 @@
-{
-  "domain": "open_router",
-  "version": "0.1",
-  "name": "OpenRouter",
-  "after_dependencies": ["assist_pipeline", "intent"],
-  "codeowners": ["@joostlek"],
-  "config_flow": true,
-  "dependencies": ["conversation"],
-  "documentation": "https://www.home-assistant.io/integrations/open_router",
-  "integration_type": "service",
-  "iot_class": "cloud_polling",
-  "quality_scale": "bronze",
-  "requirements": ["openai==2.21.0", "openrouter==0.6.0"]
-}
