+++ homeassistant/components/open_router/manifest.json
@@ -9,5 +9,5 @@
   "integration_type": "service",
   "iot_class": "cloud_polling",
   "quality_scale": "bronze",
-  "requirements": ["openai==2.21.0", "python-open-router==0.3.3"]
+  "requirements": ["openai==2.21.0", "openrouter==0.6.0"]
 }
