+++ requirements_all.txt
@@ -1695,6 +1695,9 @@ openhomedevice==2.2.0
 # homeassistant.components.openrgb
 openrgb-python==0.3.6
 
+# homeassistant.components.open_router
+openrouter==0.6.0
+
 # homeassistant.components.opensensemap
 opensensemap-api==0.2.0
 
@@ -2590,9 +2593,6 @@ python-mpd2==3.1.1
 
 # homeassistant.components.mystrom
 python-mystrom==2.5.0
-
-# homeassistant.components.open_router
-python-open-router==0.3.3
 
 # homeassistant.components.swiss_public_transport
 python-opendata-transport==0.5.0
